wallet: use SetBestBlock when saving progress during rescan
Instead of writing the best block record directly during a rescan, use
SetBestBlock. This is safe because progress is only saved in a
rescan when doing a rescan on loading. cs_wallet is held the entire time
so new blocks or reorgs will not cause the best block record to be
changed unexpectedly.