fees: document non-monotonic estimation edge case
Closes: #11800

In scenarios where data is available for higher targets but not for
lower ones, this method *may* return lower fee rates for higher
confirmation targets. This could occur if estimateCombinedFee returns no
valid data (-1) for some estimates for a low target, but **does** return
valid data for a higher target.